Ecat3.2 dev

+++ domain/src/main/java/org/fao/geonet/repository/MetadataRepositoryCustom.java
@@ -32,7 +32,9 @@ import org.fao.geonet.repository.statistic.MetadataStatisticsQueries;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -119,5 +121,5 @@ public interface MetadataRepositoryCustom {
      * 
      * @return
      */
-    Long getGaid();
+    BigInteger getGaid();
 }
